feat: onboarding agent first no clickops card
