+++ apps/dashboard/src/components/forms/status-report-update/sheet.tsx
@@ -35,8 +35,15 @@ export function FormSheetStatusReportUpdate({
   onOpenChange?: (open: boolean) => void;
 }) {
   const [internalOpen, setInternalOpen] = useState(false);
-  const open = controlledOpen ?? internalOpen;
-  const setOpen = onOpenChange ?? setInternalOpen;
+  const isControlled = controlledOpen !== undefined;
+  const open = isControlled ? controlledOpen : internalOpen;
+  const setOpen = (next: boolean) => {
+    if (isControlled) {
+      onOpenChange?.(next);
+    } else {
+      setInternalOpen(next);
+    }
+  };
   return (
     <FormSheetWithDirtyProtection open={open} onOpenChange={setOpen}>
       {children ? (
